+++ titus-common/src/test/java/com/netflix/titus/common/util/jackson/internal/ProtobufMessageObjectMapperTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Netflix, Inc.
+ * Copyright 2020 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.netflix.titus.api.json;
+package com.netflix.titus.common.util.jackson.internal;
 
 import java.util.Arrays;
 import java.util.List;
